feat: add console information
