+++ online-store/src/global.scss
@@ -501,6 +501,7 @@ body {
         display: flex;
         flex-wrap: wrap;
         justify-content: space-between;
+        gap: 20px;
 
         border-top: 1px solid #a4a4a4;
